+++ packages/moonstone/Item/Item.js
@@ -6,7 +6,8 @@
 
 import {forProp, forward, handle} from '@enact/core/handle';
 import kind from '@enact/core/kind';
-import React, {PropTypes} from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 import Slottable from '@enact/ui/Slottable';
 import Spottable from '@enact/spotlight/Spottable';