+++ packages/moonstone/VideoPlayer/VideoPlayer.js
@@ -6,6 +6,7 @@
  * @module moonstone/VideoPlayer
  */
 import React from 'react';
+import PropTypes from 'prop-types';
 import DurationFmt from '@enact/i18n/ilib/lib/DurationFmt';
 import {forward} from '@enact/core/handle';
 import ilib from '@enact/i18n';
@@ -146,7 +147,17 @@ const VideoPlayerBase = class extends React.Component {
 		 * @default 7000
 		 * @public
 		 */
-		autoCloseTimeout: React.PropTypes.number,
+		autoCloseTimeout: PropTypes.number,
+
+		/**
+		 * A string which is sent to the `backward` icon of the player controls. This can be
+		 * anything that is accepted by {@link moonstone/Icon}.
+		 *
+		 * @type {String}
+		 * @default 'backward'
+		 * @public
+		 */
+		backwardIcon: PropTypes.string,
 
 		/**
 		 * Amount of time (in milliseconds) after which the feedback text/icon part of the slider's
@@ -157,7 +168,17 @@ const VideoPlayerBase = class extends React.Component {
 		 * @default 2000
 		 * @public
 		 */
-		feedbackHideDelay: React.PropTypes.number,
+		feedbackHideDelay: PropTypes.number,
+
+		/**
+		 * A string which is sent to the `forward` icon of the player controls. This can be anything
+		 * that is accepted by {@link moonstone/Icon}.
+		 *
+		 * @type {String}
+		 * @default 'forward'
+		 * @public
+		 */
+		forwardIcon: PropTypes.string,
 
 		/**
 		 * These components are placed into the slot to the left of the media controls.
@@ -165,7 +186,25 @@ const VideoPlayerBase = class extends React.Component {
 		 * @type {Node}
 		 * @public
 		 */
-		infoComponents: React.PropTypes.node,
+		infoComponents: PropTypes.node,
+
+		/**
+		 * A string which is sent to the `jumpBackward` icon of the player controls. This can be
+		 * anything that is accepted by {@link moonstone/Icon}.
+		 *
+		 * @type {String}
+		 * @default 'skipbackward'
+		 * @public
+		 */
+		jumpBackwardIcon: PropTypes.string,
+
+		/**
+		 * Sets the `disabled` state on the media "jump" buttons; the outer pair.
+		 *
+		 * @type {Boolean}
+		 * @public
+		 */
+		jumpButtonsDisabled: PropTypes.bool,
 
 		/**
 		 * The amount of seconds the player should skip forward or backward when a "jump" button is
@@ -175,7 +214,17 @@ const VideoPlayerBase = class extends React.Component {
 		 * @default 30
 		 * @public
 		 */
-		jumpBy: React.PropTypes.number,
+		jumpBy: PropTypes.number,
+
+		/**
+		 * A string which is sent to the `jumpForward` icon of the play controls. This can be
+		 * anything that is accepted by {@link moonstone/Icon}.
+		 *
+		 * @type {String}
+		 * @default 'skipforward'
+		 * @public
+		 */
+		jumpForwardIcon: PropTypes.string,
 
 		/**
 		 * These components are placed below the title. Typically these will be media descriptor
@@ -186,7 +235,7 @@ const VideoPlayerBase = class extends React.Component {
 		 * @type {Node}
 		 * @public
 		 */
-		leftComponents: React.PropTypes.node,
+		leftComponents: PropTypes.node,
 
 		/**
 		 * Disable audio for this video. In a TV context, this is handled by the remote control,
@@ -196,7 +245,7 @@ const VideoPlayerBase = class extends React.Component {
 		 * @default false
 		 * @public
 		 */
-		muted: React.PropTypes.bool,
+		muted: PropTypes.bool,
 
 		/**
 		 * By default, the video will start playing immediately after it's loaded, unless this is set.
@@ -205,7 +254,7 @@ const VideoPlayerBase = class extends React.Component {
 		 * @default false
 		 * @public
 		 */
-		noAutoPlay: React.PropTypes.bool,
+		noAutoPlay: PropTypes.bool,
 
 		/**
 		 * Removes the "jump" buttons. The buttons that skip forward or backward in the video.
@@ -214,7 +263,7 @@ const VideoPlayerBase = class extends React.Component {
 		 * @default false
 		 * @public
 		 */
-		noJumpButtons: React.PropTypes.bool,
+		noJumpButtons: PropTypes.bool,
 
 		/**
 		 * Removes the "rate" buttons. The buttons that change the playback rate of the video.
@@ -224,7 +273,7 @@ const VideoPlayerBase = class extends React.Component {
 		 * @default false
 		 * @public
 		 */
-		noRateButtons: React.PropTypes.bool,
+		noRateButtons: PropTypes.bool,
 
 		/**
 		 * Removes the media slider.
@@ -233,7 +282,7 @@ const VideoPlayerBase = class extends React.Component {
 		 * @default false
 		 * @public
 		 */
-		noSlider: React.PropTypes.bool,
+		noSlider: PropTypes.bool,
 
 		/**
 		 * Function executed when the user clicks the Backward button. Is passed
@@ -242,7 +291,7 @@ const VideoPlayerBase = class extends React.Component {
 		 * @type {Function}
 		 * @public
 		 */
-		onBackwardButtonClick: React.PropTypes.func,
+		onBackwardButtonClick: PropTypes.func,
 
 		/**
 		 * Function executed when the user clicks the Forward button. Is passed
@@ -251,7 +300,7 @@ const VideoPlayerBase = class extends React.Component {
 		 * @type {Function}
 		 * @public
 		 */
-		onForwardButtonClick: React.PropTypes.func,
+		onForwardButtonClick: PropTypes.func,
 
 		/**
 		 * Function executed when the user clicks the JumpBackward button. Is passed
@@ -260,7 +309,7 @@ const VideoPlayerBase = class extends React.Component {
 		 * @type {Function}
 		 * @public
 		 */
-		onJumpBackwardButtonClick: React.PropTypes.func,
+		onJumpBackwardButtonClick: PropTypes.func,
 
 		/**
 		 * Function executed when the user clicks the JumpForward button. Is passed
@@ -269,7 +318,7 @@ const VideoPlayerBase = class extends React.Component {
 		 * @type {Function}
 		 * @public
 		 */
-		onJumpForwardButtonClick: React.PropTypes.func,
+		onJumpForwardButtonClick: PropTypes.func,
 
 		/**
 		 * Function executed when the user clicks the Play button. Is passed
@@ -278,7 +327,35 @@ const VideoPlayerBase = class extends React.Component {
 		 * @type {Function}
 		 * @public
 		 */
-		onPlayButtonClick: React.PropTypes.func,
+		onPlayButtonClick: PropTypes.func,
+
+		/**
+		 * A string which is sent to the `pause` icon of the player controls. This can be anything
+		 * that is accepted by {@link moonstone/Icon}.
+		 *
+		 * @type {String}
+		 * @default 'pause'
+		 * @public
+		 */
+		pauseIcon: PropTypes.string,
+
+		/**
+		 * A string which is sent to the `play` icon of the player controls. This can be anything
+		 * that is accepted by {@link moonstone/Icon}.
+		 *
+		 * @type {String}
+		 * @default 'play'
+		 * @public
+		 */
+		playIcon: PropTypes.string,
+
+		/**
+		 * Sets the `disabled` state on the media playback-rate control buttons; the inner pair.
+		 *
+		 * @type {Boolean}
+		 * @public
+		 */
+		rateButtonsDisabled: PropTypes.bool,
 
 		/**
 		 * These components are placed into the slot to the right of the media controls.
@@ -286,7 +363,17 @@ const VideoPlayerBase = class extends React.Component {
 		 * @type {Node}
 		 * @public
 		 */
-		rightComponents: React.PropTypes.node,
+		rightComponents: PropTypes.node,
+
+		/**
+		 * Any children `<source>` tag elements of [VideoPlayer]{@link moonstone/VideoPlayer} will
+		 * be sent directly to the `<video>` element as video sources.
+		 * See: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/source
+		 *
+		 * @type {Node}
+		 * @public
+		 */
+		source: PropTypes.node,
 
 		/**
 		 * Set a title for the video being played.
@@ -294,7 +381,7 @@ const VideoPlayerBase = class extends React.Component {
 		 * @type {String}
 		 * @public
 		 */
-		title: React.PropTypes.string,
+		title: PropTypes.string,
 
 		/**
 		 * The amount of time in miliseconds that should pass before the title disappears from the
@@ -304,18 +391,24 @@ const VideoPlayerBase = class extends React.Component {
 		 * @default 4000
 		 * @public
 		 */
-		titleHideDelay: React.PropTypes.number
+		titleHideDelay: PropTypes.number
 	}
 
 	static defaultProps = {
 		autoCloseTimeout: 7000,
+		backwardIcon: 'backward',
 		feedbackHideDelay: 2000,
+		forwardIcon: 'forward',
+		jumpBackwardIcon: 'skipbackward',
 		jumpBy: 30,
+		jumpForwardIcon: 'skipforward',
 		muted: false,
 		noAutoPlay: false,
 		noJumpButtons: false,
 		noRateButtons: false,
 		noSlider: false,
+		pauseIcon: 'pause',
+		playIcon: 'play',
 		titleHideDelay: 4000
 	}
 
@@ -366,7 +459,6 @@ const VideoPlayerBase = class extends React.Component {
 			more: false,
 			proportionLoaded: 0,
 			proportionPlayed: 0,
-			playPauseIcon: 'play',
 			sliderScrubbing: false,
 			sliderKnobProportion: 0,
 			titleVisible: true
@@ -550,7 +642,6 @@ const VideoPlayerBase = class extends React.Component {
 			duration: el.duration,
 			buffered: el.buffered,
 			paused: el.paused,
-			playPauseIcon: (el.paused ? 'play' : 'pause'),
 			muted: el.muted,
 			volume: el.volume,
 			playbackRate: el.playbackRate,
@@ -987,7 +1078,7 @@ const VideoPlayerBase = class extends React.Component {
 	}
 
 	render () {
-		const {children, className, infoComponents, leftComponents, noAutoPlay, noJumpButtons, noRateButtons, noSlider, rightComponents, style, title, ...rest} = this.props;
+		const {backwardIcon, children, className, forwardIcon, infoComponents, jumpBackwardIcon, jumpButtonsDisabled, jumpForwardIcon, leftComponents, noAutoPlay, noJumpButtons, noRateButtons, noSlider, pauseIcon, playIcon, rateButtonsDisabled, rightComponents, source, style, title, ...rest} = this.props;
 		delete rest.autoCloseTimeout;
 		delete rest.feedbackHideDelay;
 		delete rest.jumpBy;
@@ -1017,7 +1108,7 @@ const VideoPlayerBase = class extends React.Component {
 					ref={this.setVideoRef}
 					{...this.handledMediaEvents}
 				>
-					{children}
+					{source}
 				</video>
 
 				<Overlay onClick={this.onVideoClick}>
@@ -1055,6 +1146,11 @@ const VideoPlayerBase = class extends React.Component {
 						</MediaSlider>}
 
 						<MediaControls
+							backwardIcon={backwardIcon}
+							forwardIcon={forwardIcon}
+							jumpBackwardIcon={jumpBackwardIcon}
+							jumpButtonsDisabled={jumpButtonsDisabled}
+							jumpForwardIcon={jumpForwardIcon}
 							leftComponents={leftComponents}
 							mediaDisabled={this.state.mediaControlsDisabled}
 							moreDisabled={moreDisabled}
@@ -1065,12 +1161,15 @@ const VideoPlayerBase = class extends React.Component {
 							onForwardButtonClick={this.onForward}
 							onJumpBackwardButtonClick={this.onJumpBackward}
 							onJumpForwardButtonClick={this.onJumpForward}
+							onKeyDown={this.handleKeyDownFromControls}
 							onPlayButtonClick={this.onPlay}
 							onToggleMore={this.onMoreClick}
-							playPauseIcon={this.state.playPauseIcon}
+							paused={this.state.paused}
+							pauseIcon={pauseIcon}
+							playIcon={playIcon}
+							rateButtonsDisabled={rateButtonsDisabled}
 							rightComponents={rightComponents}
 							showMoreComponents={this.state.more}
-							onKeyDown={this.handleKeyDownFromControls}
 						>
 							{children}
 						</MediaControls>
@@ -1113,7 +1212,7 @@ const VideoPlayerBase = class extends React.Component {
  * @ui
  * @public
  */
-const VideoPlayer = Slottable({slots: ['infoComponents', 'leftComponents', 'rightComponents']}, VideoPlayerBase);
+const VideoPlayer = Slottable({slots: ['infoComponents', 'leftComponents', 'rightComponents', 'source']}, VideoPlayerBase);
 
 export default VideoPlayer;
 export {VideoPlayer, VideoPlayerBase};