+++ packages/moonstone/CHANGELOG.md
@@ -2,6 +2,28 @@
 
 The following is a curated list of changes in the Enact moonstone module, newest changes on the top.
 
+## [unreleased]
+
+### Added
+
+- `moonstone/VideoPlayer` props: `backwardIcon`, `forwardIcon`, `jumpBackwardIcon`, `jumpForwardIcon`, `pauseIcon`, and `playIcon` to support icon customization of the player
+- `moonstone/VideoPlayer` props `jumpButtonsDisabled` and `rateButtonsDisabled` for disabling the pairs of buttons when it's inappropriate for the playing media
+
+### Changed
+
+- `moonstone/Slider` colors and sizing to match the latest designs
+- `moonstone/ProgressBar` to position correctly with other components nearby
+- `moonstone/Panels` breadcrumb to no longer have a horizontal line above it
+- `moonstone/Transition` now measures itself when the CPU is idle.
+
+### Fixed
+
+- `moonstone/Popup.Popup` and `moonstone/ContextualPopupDecorator.ContextualPopupDecorator` 5-way navigation behavior
+- `moonstone/Input` to not spot its own input decorator on 5-way out
+- `moonstone/VideoPlayer` to no longer render its `children` in multiple places
+- `moonstone/Button` text color when used on a neutral (light) background in some cases
+- `moonstone/Popup` background opacity
+
 ## [1.0.0] - 2017-03-31
 
 > NOTE: We have also modified most form components to be usable in a controlled (app manages component